--- linkup/src/main/java/com/learningcrew/linkup/common/constants/Gender.java
@@ -1,5 +1,0 @@
-package com.learningcrew.linkup.common.constants;
-
-public enum Gender {
-    M, F, BOTH
-}
