--- linkup/src/main/java/com/learningcrew/linkup/common/constants/IsActive.java
@@ -1,5 +1,0 @@
-package com.learningcrew.linkup.common.constants;
-
-public enum IsActive {
-    Y, N
-}
