--- linkup/src/main/java/com/learningcrew/linkup/common/constants/StatusType.java
@@ -1,5 +1,0 @@
-package com.learningcrew.linkup.common.constants;
-
-public enum StatusType {
-    PENDING, ACCEPTED, REJECTED, DELETED;
-}
