--- linkup/src/main/java/com/learningcrew/linkup/community/query/dto/response/dummy
@@ -1,1 +1,0 @@
-
