--- linkup/src/main/java/com/learningcrew/linkup/community/query/mapper/dummy
@@ -1,1 +1,0 @@
-
