--- linkup/src/main/java/com/learningcrew/linkup/community/query/service/dummy
@@ -1,1 +1,0 @@
-
