--- linkup/src/main/java/com/learningcrew/linkup/linker/command/domain/aggregate/Role.java
@@ -1,17 +1,0 @@
-package com.learningcrew.linkup.linker.command.domain.aggregate;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "role")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Role {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int role_id;
-    private String roleName;
-}
