--- linkup/src/main/java/com/learningcrew/linkup/meeting/command/domain/aggregate/Member.java
@@ -1,21 +1,0 @@
-package com.learningcrew.linkup.meeting.command.domain.aggregate;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.apache.ibatis.type.Alias;
-
-@Alias("MeetingMember")
-@Entity(name = "MeetingMember")
-@Table(name = "member")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Member {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int memberId;
-
-    private String nickname;
-}
