--- linkup/src/main/java/com/learningcrew/linkup/point/query/dto/response/dummy
@@ -1,1 +1,0 @@
-
