--- linkup/src/main/java/com/learningcrew/linkup/point/query/mapper/dummy
@@ -1,1 +1,0 @@
-
