--- linkup/src/main/java/com/learningcrew/linkup/point/query/service/dummy
@@ -1,1 +1,0 @@
-
