--- linkup/src/main/java/com/learningcrew/linkup/report/query/dto/response/dummy
@@ -1,1 +1,0 @@
-
