--- linkup/src/main/java/com/learningcrew/linkup/review/query/dto/response/dummy
@@ -1,1 +1,0 @@
-
