--- linkup/src/main/java/com/learningcrew/linkup/review/query/mapper/dummy
@@ -1,1 +1,0 @@
-
