--- linkup/src/main/java/com/learningcrew/linkup/review/query/service/dummy
@@ -1,1 +1,0 @@
-
